(second stage) Added bias_subtractor and Exp_adder
These two component work to give out the sum of the two exponent with the bias already subtracted.
The exponent adder controls that both of the exp aren't zero. If one of them is, it sets it to -126.  The sum is done with a CLA, made with 8 Full Adders.
The bias subtractor adds a 0 to the MSB of both exponents, after that it negates the bias to get it in module 1. It uses a CLA  made with 8 Full Adders on the first 8 lsb, and a CLA on the last 2 msb, which he gets from the first CLA.