Added final components
These final components are the represantive of the final stage, the rounder finds the first 1 in the 23 msb of the mantix, and gets the offset of it, which is the number we have to shift the mantix to the left to fix it. If it's all zero then we get the final 23 bits. From there we calculate the final exponent value, with the offset value and the intermediate_exp. The rounder finds out if we have an overflow, underflow, normalized or denormalized number. After that we check if we have eventually some flags up for the invalid value, and we give the output.